requirements.txt
# =========================
# Core runtime
# =========================
python-dotenv>=1.0.1

# =========================
# MCP / FastMCP
# =========================
fastmcp>=0.4.0
httpx>=0.27.0
pydantic>=2.6.0

# =========================
# Scheduling & orchestration
# =========================
APScheduler>=3.10.4

# =========================
# Database / persistence
# =========================
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25

# =========================
# Vector database
# =========================
chromadb>=0.4.24

# =========================
# LLM / AI stack
# =========================
openai>=1.12.0
tiktoken>=0.6.0

# =========================
# Data & forecasting
# =========================
numpy>=1.26.3
pandas>=2.2.0
scikit-learn>=1.4.0

# =========================
# Security / crypto
# =========================
cryptography>=42.0.2
passlib[bcrypt]>=1.7.4
python-jose>=3.3.0

# =========================
# Utilities
# =========================
rich>=13.7.0
tenacity>=8.2.3

# =========================
# Testing
# =========================
pytest>=8.0.0
